+++ charts/aws-fsx-csi-driver/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Helm chart
 
+# v1.13.0
+* Use driver image 1.6.0
+
+# v1.12.1
+* Parameterize controller.extraTags
+
+# v1.12.0
+* Use driver image 1.5.0
+
 # v1.11.0
 * Use driver image 1.4.0